// SGUEES-SPA/src/app/pages/Shop/com-bitacora/models/com-bitacora.ts
export interface ComBitacora {
	CORR_EMPRESA: number;
	CORR_BITACORA: number;
	FECHA_BITACORA: Date;
	CODIGO_OPCION: string;
	NOMBRE_CODIGO_OPCION: string;
	CLASE_BITACORA: string;
	NOMBRE_CLASE_BITACORA: string;
	LLAVE_TRANSACCION: string;
	REFERENCIA_TRANSACCION: string;
	USUARIO_CREA_TRANS: string;
	ESTACION_CREA_TRANS: string;
	FECHA_CREA_TRANS: Date;
	USUARIO_ACTU_TRANS: string;
	ESTACION_ACTU_TRANS: string;
	FECHA_ACTU_TRANS: Date;
	USUARIO_CREA_BITACORA: string;
	ESTACION_CREA_BITACORA: string;
	FECHA_CREA_BITACORA: Date;
}